Remove nidx_vector service layer
Remove the service layer from the vector index:
- The reader had logic parsing the search request, this has been moved
  to the data point provider reader, including tests.
- The writer has been deleted. The set_resource method has been
  extracted to indexer.rs as index_resource.
