+++ nidx/nidx_vector/tests/test_hidden.rs
@@ -22,8 +22,10 @@ use std::collections::{HashMap, HashSet};
 
 use nidx_protos::{IndexParagraph, IndexParagraphs, Resource, ResourceId, VectorSearchRequest, VectorSentence};
 use nidx_vector::config::VectorConfig;
+use nidx_vector::data_point_provider::reader::{Reader, VectorsContext};
+use nidx_vector::data_point_provider::writer::Writer;
+use nidx_vector::indexer::{index_resource, ResourceWrapper};
 use nidx_vector::query_language::BooleanExpression;
-use nidx_vector::service::{ResourceWrapper, VectorReaderService, VectorWriterService, VectorsContext};
 use tempfile::tempdir;
 use uuid::Uuid;
 
@@ -63,18 +65,29 @@ fn resource(labels: Vec<String>) -> Resource {
 fn test_hidden_search() -> anyhow::Result<()> {
     let workdir = tempdir()?;
     let index_path = workdir.path().join("vectors");
-    let mut writer = VectorWriterService::create(&index_path, VectorConfig::default())?;
+    let mut writer = Writer::new(&index_path, VectorConfig::default())?;
 
     // Create two resources, one hidden and one not
     let labels = vec!["/q/h".to_string()];
     let hidden_resource = resource(labels);
-    writer.set_resource(ResourceWrapper::from(&hidden_resource))?;
+    let (Some(pin), _deletions) =
+        index_resource(ResourceWrapper::from(&hidden_resource), &index_path, writer.config())?
+    else {
+        panic!("No segment indexed")
+    };
+    writer.add_data_point(pin)?;
 
     let visible_resource = resource(vec![]);
-    writer.set_resource(ResourceWrapper::from(&visible_resource))?;
+    let (Some(pin), _deletions) =
+        index_resource(ResourceWrapper::from(&visible_resource), &index_path, writer.config())?
+    else {
+        panic!("No segment indexed")
+    };
+    writer.add_data_point(pin)?;
+    writer.commit()?;
 
     // Find all resources
-    let reader = VectorReaderService::open(&index_path)?;
+    let reader = Reader::open(&index_path)?;
     let request = VectorSearchRequest {
         vector: vec![0.5, 0.5, 0.5, 0.5],
         min_score: -1.0,
